+++ CHANGELOG.md
@@ -7,5 +7,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - Unreleased
+### Added
+- `url` attribute to load the json automatically.
+- `label` attribute to insert a label for the input.
+- `placeholder` attribute to configure the placeholder of the input.
+- Improved styling.
+
+### Changed
+- Simplify the code to only one component.
+- The component is not auto-registered.
+
 ## 1.0.0 - 2021-07-31
 First version
+
+[2.0.0]: https://github.com/oom-components/searcher/compare/v1.0.0...HEAD